Replace gulp build with webpack 4
