--- .editorconfig
@@ -1,16 +1,0 @@
-# EditorConfig is awesome: http://EditorConfig.org
-
-root = true;
-
-[*]
-#  Ensure there's no lingering whitespace
-trim_trailing_whitespace = true
-# Ensure a newline at the end of each file
-insert_final_newline = true
-
-[*.js]
-# Unix-style newlines
-end_of_line = lf
-charset = utf-8
-indent_style = space
-indent_size = 2
