+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+### 1.0.0
+
+- Replaced gulp build with webpack 4
+- Rename repository to `magma-script-js`
+- Rename file `MagmaScript.js` to `magma-script-evaluator.js`
+- Merge test files due to mocha-webpack not running wildcards
+- Set moment and mathjs to external dependencies, reducing library size
+- Rename library to @molgenis/magma-script-js
+
 ### 0.0.9
 
 - Move Molgenis MagmaScript evaluator to npm module
