+++ package.json
@@ -1,67 +1,37 @@
 {
-  "name": "@molgenis/magma-script",
-  "version": "0.0.9",
+  "name": "@molgenis/magma-script-js",
+  "version": "1.0.0",
   "description": "MagmaScript evaluator in JavaScript",
-  "main": "dist/MagmaScript.js",
-  "files": [
-    "dist"
-  ],
-  "unpkg": "dist/MagmaScript.min.js",
-  "scripts": {
-    "test": "gulp",
-    "lint": "gulp lint",
-    "test-browser": "gulp test-browser",
-    "watch": "gulp watch",
-    "build": "gulp build",
-    "coverage": "gulp coverage"
-  },
   "repository": {
     "type": "git",
-    "url": "https://github.com/molgenis/magma-script.git"
+    "url": "https://github.com/molgenis/magma-script-js.git"
+  },
+  "author": "MOLGENIS Team, System Genetics, UMCG, Netherlands",
+  "license": "LGPL-3.0",
+  "bugs": {
+    "url": "https://github.com/molgenis/magma-script-js/issues"
   },
   "keywords": [
+    "molgenis",
     "MagmaScript"
   ],
-  "author": "Fleur Kelpin <f.kelpin@umcg.nl>",
-  "license": "LGPL-3.0+",
-  "bugs": {
-    "url": "https://github.com/molgenis/magma-script/issues"
+  "main": "dist/magma-script-evaluator.js",
+  "module": "src/magma-script-evaluator.js",
+  "scripts": {
+    "build": "webpack --mode production",
+    "test": "mocha-webpack --mode production \"test/unit/magma-script-evaluator.js.spec.js\""
   },
-  "homepage": "https://github.com/molgenis/magma-script",
-  "devDependencies": {
-    "babel-core": "6.14.0",
-    "babel-loader": "6.2.5",
-    "babel-polyfill": "6.13.0",
-    "babel-preset-latest": "6.14.0",
-    "babel-register": "6.14.0",
-    "chai": "3.5.0",
-    "del": "2.2.2",
-    "glob": "7.0.6",
-    "gulp": "3.9.1",
-    "gulp-eslint": "3.0.1",
-    "gulp-filter": "4.0.0",
-    "gulp-istanbul": "1.1.1",
-    "gulp-livereload": "3.8.1",
-    "gulp-load-plugins": "1.2.4",
-    "gulp-mocha": "3.0.1",
-    "gulp-plumber": "1.1.0",
-    "gulp-rename": "1.2.2",
-    "gulp-sourcemaps": "1.6.0",
-    "gulp-uglify": "2.0.0",
-    "isparta": "4.0.0",
-    "json-loader": "0.5.4",
-    "mocha": "3.0.2",
-    "sinon": "1.17.5",
-    "sinon-chai": "2.8.0",
-    "webpack": "1.13.2",
-    "webpack-stream": "3.2.0"
-  },
-  "babelBoilerplateOptions": {
-    "entryFileName": "MagmaScript.js",
-    "mainVarName": "MagmaScript"
-  },
+  "homepage": "https://github.com/molgenis/magma-script-js",
   "dependencies": {
     "mathjs": "^4.3.0",
     "moment": "^2.22.1"
+  },
+  "devDependencies": {
+    "chai": "^4.1.2",
+    "mocha": "^5.2.0",
+    "mocha-webpack": "2.0.0-beta.0",
+    "uglifyjs-webpack-plugin": "^1.2.5",
+    "webpack": "^4.9.1",
+    "webpack-cli": "^2.1.4"
   }
 }
