+++ src/magma-script-evaluator.js
