+++ test/.eslintrc
@@ -1,14 +1,4 @@
 {
-  "extends": "./setup/.globals.json",
-  "parserOptions": {
-    "ecmaVersion": 6,
-    "sourceType": "module"
-  },
-  "rules": {
-    "strict": 0,
-    "quotes": [2, "single"],
-    "no-unused-expressions": 0
-  },
   "env": {
     "browser": true,
     "node": true,
