--- test/setup/.globals.json
@@ -1,14 +1,0 @@
-{
-  "globals": {
-    "expect": true,
-    "mock": true,
-    "sandbox": true,
-    "spy": true,
-    "stub": true,
-    "useFakeServer": true,
-    "useFakeTimers": true,
-    "useFakeXMLHttpRequest": true,
-    "math": true,
-    "moment": true
-  }
-}
