--- test/setup/browser.js
@@ -1,9 +1,0 @@
-var mochaGlobals = require('./.globals.json').globals;
-
-window.mocha.setup('bdd');
-window.onload = function() {
-  window.mocha.checkLeaks();
-  window.mocha.globals(Object.keys(mochaGlobals));
-  window.mocha.run();
-  require('./setup')(window);
-};
